Use existing webhooks to determine whether messages should be modified
Don't store any data regarding which channels should be modified.
Instead, check if there are any existing webhooks for this channel.
As long as a webhook is present, messages will be modified.
This way information about which channel should be modified is not stored in the bot and will survive bot restart.